CURB-4028 rename cookie folder to tracking, adjust imports
