+++ themes/theme-gmd/components/NavDrawer/components/StoreInfo/components/PrivacySettingsButton/index.jsx
@@ -4,11 +4,11 @@ import {
   NAV_MENU_PRIVACY_SETTINGS,
   NAV_MENU_PRIVACY_SETTINGS_AFTER,
   NAV_MENU_PRIVACY_SETTINGS_BEFORE,
-} from '@shopgate/pwa-common/constants/Portals';
-import Portal from '@shopgate/pwa-common/components/Portal';
+} from '@shopgate/engage/core';
+import { Portal } from '@shopgate/engage/components';
 import SecurityIcon from '@shopgate/pwa-ui-shared/icons/SecurityIcon';
 import { NavDrawer } from '@shopgate/pwa-ui-material';
-import { COOKIE_CONSENT_PATTERN } from '@shopgate/engage/cookie-consent/constants';
+import { COOKIE_CONSENT_PATTERN } from '@shopgate/engage/tracking/constants';
 import portalProps from '../../../../portalProps';
 import connect from '../../../../connector';
