+++ themes/theme-gmd/pages/reducers.js
@@ -22,7 +22,7 @@ import product from '@shopgate/pwa-common-commerce/product/reducers';
 import search from '@shopgate/pwa-common-commerce/search/reducers';
 import reviews from '@shopgate/pwa-common-commerce/reviews/reducers';
 import extensions from 'Extensions/reducers';
-import cookieConsent from '@shopgate/engage/cookie-consent/reducers';
+import cookieConsent from '@shopgate/engage/tracking/reducers';
 
 persistedReducers.set([
   'cart',
