+++ themes/theme-ios11/pages/More/components/StoreInfo/index.jsx
@@ -4,7 +4,8 @@ import {
   NAV_MENU_STORE_INFORMATION_BEFORE,
   NAV_MENU_STORE_INFORMATION,
   NAV_MENU_STORE_INFORMATION_AFTER,
-} from '@shopgate/pwa-common/constants/Portals';
+} from '@shopgate/engage/core';
+import { appConfig } from '@shopgate/engage';
 import portalProps from '../../portalProps';
 import Section from '../Section';
 import Shipping from './components/Shipping';
@@ -19,22 +20,26 @@ import PrivacySettings from './components/PrivacySettings';
  * The StoreInfoComponent.
  * @returns {JSX.Element}
  */
-const StoreInfo = () => (
-  <Fragment>
-    <Portal name={NAV_MENU_STORE_INFORMATION_BEFORE} props={portalProps} />
-    <Portal name={NAV_MENU_STORE_INFORMATION} props={portalProps}>
-      <Section title="navigation.store_information">
-        <Shipping />
-        <Payment />
-        <Terms />
-        <Privacy />
-        <PrivacySettings />
-        <ReturnPolicy />
-        <Imprint />
-      </Section>
-    </Portal>
-    <Portal name={NAV_MENU_STORE_INFORMATION_AFTER} props={portalProps} />
-  </Fragment>
-);
+const StoreInfo = () => {
+  const { cookieConsent: { cookieConsentActivated } } = appConfig;
+
+  return (
+    <Fragment>
+      <Portal name={NAV_MENU_STORE_INFORMATION_BEFORE} props={portalProps} />
+      <Portal name={NAV_MENU_STORE_INFORMATION} props={portalProps}>
+        <Section title="navigation.store_information">
+          <Shipping />
+          <Payment />
+          <Terms />
+          <Privacy />
+          {cookieConsentActivated && <PrivacySettings />}
+          <ReturnPolicy />
+          <Imprint />
+        </Section>
+      </Portal>
+      <Portal name={NAV_MENU_STORE_INFORMATION_AFTER} props={portalProps} />
+    </Fragment>
+  );
+};
 
 export default StoreInfo;
