+++ themes/theme-ios11/pages/index.jsx
@@ -40,8 +40,8 @@ import Viewport from 'Components/Viewport';
 import Dialog from '@shopgate/pwa-ui-shared/Dialog';
 import { PushOptInModal } from '@shopgate/engage/push-opt-in/components';
 import { BACK_IN_STOCK_PATTERN } from '@shopgate/engage/back-in-stock/constants';
-import { COOKIE_CONSENT_PATTERN } from '@shopgate/engage/cookie-consent/constants';
-import { CookieConsentModal } from '@shopgate/engage/cookie-consent/components';
+import { COOKIE_CONSENT_PATTERN } from '@shopgate/engage/tracking/constants';
+import { CookieConsentModal } from '@shopgate/engage/tracking/components';
 import themeApi from '../themeApi';
 import * as routes from './routes';
 import { routesTransforms } from './routesTransforms';
