+++ themes/theme-ios11/pages/subscribers.js
@@ -47,7 +47,7 @@ import extensions from 'Extensions/subscribers';
 import engageBackInStock from '@shopgate/engage/back-in-stock/subscriptions';
 import appRating from '@shopgate/engage/app-rating/subscriptions';
 import pushOptIn from '@shopgate/engage/push-opt-in/subscriptions';
-import cookieConsent from '@shopgate/engage/cookie-consent/subscriptions/cookieConsent';
+import cookieConsent from '@shopgate/engage/tracking/subscriptions';
 
 const subscriptions = [
   // Common subscribers.
